scripts/requirements-api-excel.txt
openpyxl==3.1.5
pycryptodome==3.20.0
